Fix cloud, some cleanup

+++ hook/hook.h
@@ -3,7 +3,7 @@
 #include "avs.h"
 
 int hook_avs_fs_fstat(int flags, struct avs_stat *st);
-AVS_FILE hook_avs_fs_open(const char* name, int mode, int flags);
+AVS_FILE hook_avs_fs_open(const char* name, uint16_t mode, int flags);
 
 extern "C" {
 	__declspec(dllexport) int init(void);
